Added initial skeleton of the application.
Added the license.

Coded up the initial skeleton of the interface.

Started reviewing TPC-C 5.11 document, and making notes of important
things that may affect the implementation of the benchmark.

Added 'pg' and 'blessed' as the dependencies of this application.

Added a lot more meat to the README, explaining the need of a new TPC-C
implemetation, and specifically why I chose NodeJS based implementation.