Drive code compilation and test runs via Makefile.
This commit introduces long-overdue Makefile based builds.

Introduced 3 targets, one each to compile, run and watch for changes.

Added ': any' type specification to many variables whose data type I'm
not aware of. Added appropriate types for the ones that belong to this
project.

Added the build-artifact, runner.js, to the project for the first time.
I had been avoiding this for a long time since the build intructions
were still in flux. The new 'Developer' section in TPC-C/README.md tries
to address this, but agreeably it's still a bit lacking.